Faltan validaciones

+++ control-incidencias/src/main/java/com/is/controlincidencias/service/impl/AsistenciaServiceImpl.java
@@ -5,6 +5,7 @@ import com.is.controlincidencias.entity.Personal;
 import com.is.controlincidencias.model.AsistenciaForm;
 import com.is.controlincidencias.model.AsistenciaJSON;
 import com.is.controlincidencias.repository.AsistenciaRepository;
+import com.is.controlincidencias.repository.PeriodoInhabilRepository;
 import com.is.controlincidencias.repository.PersonalRepository;
 import com.is.controlincidencias.service.AsistenciaService;
 import lombok.extern.slf4j.Slf4j;
@@ -12,6 +13,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
 
+import java.time.DayOfWeek;
 import java.time.LocalDate;
 import java.time.LocalTime;
 import java.util.ArrayList;
@@ -22,18 +24,25 @@ import java.util.List;
 public class AsistenciaServiceImpl implements AsistenciaService {
     private static LocalTime siete = LocalTime.of(7, 0, 0, 0);
     private static LocalTime veintidos = LocalTime.of(22, 0, 0, 0);
-    private static LocalTime seisMedia = LocalTime.of(6, 30, 0);
-    private static LocalTime veintitres = LocalTime.of(23, 0, 0, 0);
+    private static LocalTime seis = LocalTime.of(6, 0, 0);
+
+    private static final int NO_EXISTE_TARJETA = 1;
+    private static final int REGISTRO_DUPLICADO = 2;
+    private static final int FECHA_ASISTENCIA_INVALIDA = 5;
 
     private final AsistenciaRepository asistenciaRepository;
     private final PersonalRepository personalRepository;
+    private final PeriodoInhabilRepository periodoInhabilRepository;
 
     @Autowired
     public AsistenciaServiceImpl(
             @Qualifier("asistenciaRepository") AsistenciaRepository asistenciaRepository,
-            @Qualifier("personalRepository") PersonalRepository personalRepository) {
+            @Qualifier("personalRepository") PersonalRepository personalRepository,
+            @Qualifier("periodoInhabilRepository") PeriodoInhabilRepository
+                    periodoInhabilRepository) {
         this.asistenciaRepository = asistenciaRepository;
         this.personalRepository = personalRepository;
+        this.periodoInhabilRepository = periodoInhabilRepository;
     }
 
     @Override
@@ -48,30 +57,44 @@ public class AsistenciaServiceImpl implements AsistenciaService {
     }
 
     @Override
-    public void agregarAsistencia(AsistenciaJSON asistenciaJSON) {
-        Personal p = personalRepository.getPersonalByNoTarjeta(asistenciaJSON.getTarjeta());
+    public Asistencia agregarAsistencia(AsistenciaForm form) {
+        Personal p = personalRepository.getPersonalByNoTarjeta(form.getTarjeta());
         Asistencia asistencia = new Asistencia();
         asistencia.setPersonal(p);
-        asistencia.setFechaRegistro(asistenciaJSON.getFecha());
-        if (asistenciaJSON.getHoraEntrada().compareTo(siete) > 0) asistencia.setHoraEntrada(
-                asistenciaJSON.getHoraEntrada());
-        else asistencia.setHoraEntrada(siete);
-        if (asistenciaJSON.getHoraSalida().compareTo(veintidos) < 0) asistencia.setHoraSalida(
-                asistenciaJSON.getHoraSalida());
-        else asistencia.setHoraSalida(veintidos);
-        Asistencia a = asistenciaRepository.save(asistencia);
-        log.info("Id Asistencia: " + a.getIdAsistencia());
+        if (!horasValidas(form.getHoraEntrada(), form.getHoraSalida(), p.getTipo()))
+            return null;
+        asistencia.setFechaRegistro(form.getFecha());
+        asistencia.setHoraEntrada(form.getHoraEntrada());
+        asistencia.setHoraSalida(form.getHoraSalida());
+
+        return asistenciaRepository.save(asistencia);
     }
 
     @Override
     public int existeAsistencia(AsistenciaForm asistenciaForm) {
-        boolean existeTarjeta = personalRepository.existsPersonalByNoTarjeta(
-                asistenciaForm.getTarjeta());
-        if (!existeTarjeta) return 1; // No existe la tarjeta
+        Personal p = personalRepository.findByNoTarjeta(asistenciaForm.getTarjeta());
+        if (p == null) return NO_EXISTE_TARJETA; // No existe la tarjeta
 
-        Asistencia a = asistenciaRepository.findAsistenciaByFechaRegistroAndPersonalNoTarjeta(
-                asistenciaForm.getFecha(), asistenciaForm.getTarjeta());
-        if (a == null) return 2; // No se encontro registro
+        LocalDate fechaFinal = LocalDate.now();
+        LocalDate fechaIncial = fechaFinal;
+        log.info("existeAsistencia() AQUI 1");
+        if (esInhabil(asistenciaForm.getFecha())) return FECHA_ASISTENCIA_INVALIDA;
+        int contador = 0;
+        while (contador < 2) {
+            fechaIncial = fechaIncial.minusDays(1);
+            if (!esInhabil(fechaIncial)) {
+                contador++;
+            }
+        }
+        log.info("existeAsistencia() fechaInicial =" + fechaIncial);
+        log.info("existeAsistencia() fechaFinal =" + fechaFinal);
+        if (asistenciaForm.getFecha().isAfter(fechaFinal) || asistenciaForm.getFecha().isBefore(
+                fechaIncial)) return FECHA_ASISTENCIA_INVALIDA;
+
+        boolean existe = asistenciaRepository
+                .existsAsistenciaByFechaRegistroAndPersonalNoTarjeta(asistenciaForm.getFecha(),
+                        asistenciaForm.getTarjeta());
+        if (existe) return REGISTRO_DUPLICADO;
 
         return 0; // Salio chido
     }
@@ -80,32 +103,32 @@ public class AsistenciaServiceImpl implements AsistenciaService {
     public AsistenciaForm buscarAsistencia(AsistenciaForm asistenciaForm) {
         Asistencia a = asistenciaRepository.findAsistenciaByFechaRegistroAndPersonalNoTarjeta(
                 asistenciaForm.getFecha(), asistenciaForm.getTarjeta());
+        Personal p = personalRepository.getPersonalByNoTarjeta(asistenciaForm.getTarjeta());
 
-        asistenciaForm.setHoraSalida(a.getHoraSalida());
-        asistenciaForm.setHoraEntrada(a.getHoraEntrada());
         asistenciaForm.setFecha(asistenciaForm.getFecha());
         asistenciaForm.setTarjeta(asistenciaForm.getTarjeta());
         asistenciaForm.setId(asistenciaForm.getId());
-        asistenciaForm.setNombre(a.getPersonal().getNombre() + " "
-                + a.getPersonal().getApellidoPaterno() + " " + " "
-                + a.getPersonal().getApellidoMaterno());
+        asistenciaForm.setNombre(p.getNombre() + " " + p.getApellidoPaterno() + " " + p.getApellidoMaterno());
+        if (a != null) {
+            asistenciaForm.setHoraSalida(a.getHoraSalida());
+            asistenciaForm.setHoraEntrada(a.getHoraEntrada());
 
+        }
         return asistenciaForm;
     }
 
     @Override
-    public Asistencia modificarAsistencia(AsistenciaForm asistenciaForm) {
+    public Asistencia modificarAsistencia(AsistenciaForm form) {
+        Personal p = personalRepository.findByNoTarjeta(form.getTarjeta());
+        Asistencia a = asistenciaRepository.findByIdAsistencia(form.getId());
 
-        LocalTime horaEntrada = asistenciaForm.getHoraEntrada();
-        LocalTime horaSalida = asistenciaForm.getHoraSalida();
-        Asistencia a = asistenciaRepository.findByIdAsistencia(asistenciaForm.getId());
-        if (a == null) return null;
-        if (!validarHoras(horaEntrada, horaSalida)) return null;
+        if (!horasValidas(form.getHoraEntrada(), form.getHoraSalida(), p.getTipo()))
+            return null;
+        else
+            log.info("modificarAsistencia() horas validas");
 
-        if (horaEntrada.compareTo(siete) > 0) a.setHoraEntrada(horaEntrada);
-        else a.setHoraEntrada(siete);
-        if (horaSalida.compareTo(veintidos) < 0) a.setHoraSalida(horaSalida);
-        else a.setHoraSalida(veintidos);
+        a.setHoraEntrada(form.getHoraEntrada());
+        a.setHoraSalida(form.getHoraSalida());
 
         return asistenciaRepository.save(a);
     }
@@ -127,13 +150,33 @@ public class AsistenciaServiceImpl implements AsistenciaService {
 
     @Override
     public void eliminarAsistenciaPorId(Integer id) {
-        if (asistenciaRepository.existsById(id))
-            asistenciaRepository.deleteById(id);
+        if (asistenciaRepository.existsById(id)) asistenciaRepository.deleteById(id);
     }
 
-    private boolean validarHoras(LocalTime horaEntrada, LocalTime horaSalida) {
-        if (horaEntrada.compareTo(seisMedia) < 0 || horaEntrada.compareTo(veintitres) > 0)
-            return false;
-        return horaSalida.compareTo(seisMedia) >= 0 && horaSalida.compareTo(veintitres) <= 0;
+    private boolean horasValidas(LocalTime entrada, LocalTime salida, String rol) {
+        if (rol.equals("ROLE_PAAE")) {
+            if (entrada.isBefore(seis) || salida.isAfter(veintidos) || entrada.isAfter(salida))
+                return false;
+            log.info("ES PAAE");
+        } else {
+            if (entrada.isBefore(siete) || salida.isAfter(veintidos) || entrada.isAfter(salida))
+                return false;
+            log.info("NO ES PAAE");
+        }
+
+        log.info("horasValidas() es true :" + entrada.isBefore(siete) + " " + salida.isAfter
+                (veintidos) + " " + entrada.isAfter(salida));
+        return true;
+    }
+
+    private boolean esInhabil(LocalDate fecha) {
+        boolean inhabil = periodoInhabilRepository
+                .existsByInicioIsLessThanEqualAndFinGreaterThanEqual(fecha, fecha);
+        log.info("esInhabil() 1= " + inhabil);
+        if (fecha.getDayOfWeek() == DayOfWeek.SATURDAY || fecha.getDayOfWeek() == DayOfWeek.SUNDAY) {
+            inhabil = true;
+        }
+        log.info("esInhabil() 2= " + inhabil);
+        return inhabil;
     }
 }
